restory withqueryloader test
