+++ components/AlertsDisplay.tsx
@@ -1,20 +1,23 @@
-import React from 'react'
+import React, { useState, useEffect } from 'react'
 import Alert from './Alert'
-import { AlertData } from '../@types/alerts'
+import { AlertData, DismissedAlerts } from '../@types/alerts'
 
 type Props = {
   alerts: AlertData[]
-  dismissedAlerts: { [id: string]: boolean }
-  setDismissedAlerts: React.Dispatch<React.SetStateAction<{}>>
   page?: string
 }
 
-const AlertsDisplay: React.FC<Props> = ({
-  alerts,
-  dismissedAlerts,
-  setDismissedAlerts,
-  page
-}) => {
+const AlertsDisplay: React.FC<Props> = ({ alerts, page }) => {
+  const [dismissedAlerts, setDismissedAlerts] = useState<DismissedAlerts>({})
+  useEffect(() => {
+    const localDismissedAlerts = localStorage.getItem('dismissedAlerts')
+    if (localDismissedAlerts) {
+      setDismissedAlerts(JSON.parse(localDismissedAlerts))
+    }
+  }, [])
+  useEffect(() => {
+    localStorage.setItem('dismissedAlerts', JSON.stringify(dismissedAlerts))
+  }, [dismissedAlerts])
   const widthClass = page === 'curriculum' ? 'col-12' : ''
   return (
     <div className={`alerts-container ${widthClass}`}>
