+++ containers/withQueryLoader.test.js
@@ -14,25 +14,18 @@ describe('withQueryLoader HOC container', () => {
   })
 
   test('Should return Component when not loading and data is present', () => {
-    useQuery.mockReturnValue({ loading: false, data: { a: 1, alerts: [] } })
+    useQuery.mockReturnValue({ loading: false, data: { a: 1 } })
     const Component = () => {
       return <h1>Component</h1>
     }
 
     expect(
       withQueryLoader({ query: null }, Component)({ title: 'test' })
-    ).toEqual(
-      <Component
-        queryData={{ a: 1, alerts: [] }}
-        dismissedAlerts={[]}
-        setDismissedAlerts={jest.fn()}
-        title="test"
-      />
-    )
+    ).toEqual(<Component queryData={{ a: 1 }} title="test" />)
   })
 
   test('Should return No data when done loading and no data is present', () => {
-    useQuery.mockReturnValue({ loading: false, data: null })
+    useQuery.mockReturnValue({ loading: false })
     const Component = () => {
       return <h1>Component</h1>
     }
