+++ containers/withQueryLoader.tsx
@@ -1,8 +1,7 @@
-import React, { useEffect, useState } from 'react'
+import React from 'react'
 import { useQuery } from '@apollo/react-hooks'
 import { DocumentNode } from 'apollo-boost'
 import LoadingSpinner from '../components/LoadingSpinner'
-import { DismissedAlerts } from '../@types/alerts'
 
 type QueryProps = {
   query: DocumentNode
@@ -11,8 +10,6 @@ type QueryProps = {
 
 export type WithQueryProps = {
   queryData: any
-  dismissedAlerts: DismissedAlerts
-  setDismissedAlerts: React.Dispatch<React.SetStateAction<DismissedAlerts>>
 }
 
 const withQueryLoader = (
@@ -20,28 +17,11 @@ const withQueryLoader = (
   Component: React.FC<any>
 ) => (props: any) => {
   const { loading, data } = useQuery(query, getParams(props))
-  const [dismissedAlerts, setDismissedAlerts] = useState<DismissedAlerts>({})
-  useEffect(() => {
-    const localDismissedAlerts = localStorage.getItem('dismissedAlerts')
-    if (localDismissedAlerts) {
-      setDismissedAlerts(JSON.parse(localDismissedAlerts))
-    }
-  }, [])
-  useEffect(() => {
-    localStorage.setItem('dismissedAlerts', JSON.stringify(dismissedAlerts))
-  }, [dismissedAlerts])
   if (loading) {
     return <LoadingSpinner />
   }
   if (data) {
-    return (
-      <Component
-        queryData={data}
-        dismissedAlerts={dismissedAlerts}
-        setDismissedAlerts={setDismissedAlerts}
-        {...props}
-      />
-    )
+    return <Component queryData={data} {...props} />
   }
 
   return <h1>No Data...</h1>
