+++ pages/curriculum.tsx
@@ -10,21 +10,17 @@ import withQueryLoader, { WithQueryProps } from '../containers/withQueryLoader'
 import AlertsDisplay from '../components/AlertsDisplay'
 import _ from 'lodash'
 
-export const Curriculum: React.FC<WithQueryProps> = ({
-  queryData,
-  dismissedAlerts,
-  setDismissedAlerts
-}) => {
+type LessonStatusMap = {
+  [id: string]: LessonStatus
+}
+
+export const Curriculum: React.FC<WithQueryProps> = ({ queryData }) => {
   const announcements = [
     'To make space for other students on our servers, your account will be deleted after 30 days of inactivity.',
     'Take each lesson challenge seriously and do them over and over again until you can solve them. With the exception End to End, all challenges are questions and exercises taken from real interviews.',
     'This lesson will not only prepare you for interviews, but it will also help teach you the skills that you need to become an effective engineer.',
     'After completing Foundations of JavaScript, Variables & Functions, Array, Objects, End to End, HTML/CSS/JavaScript, React/GraphQL/SocketIO, you will be technically ready to contribute to our codebase.'
   ]
-
-  type LessonStatusMap = {
-    [id: string]: LessonStatus
-  }
 
   const { lessons, session, alerts } = queryData
   const lessonStatus: LessonStatus[] = _.get(session, 'lessonStatus', [])
@@ -78,12 +74,7 @@ export const Curriculum: React.FC<WithQueryProps> = ({
   return (
     <Layout>
       <div className="row">
-        <AlertsDisplay
-          alerts={alerts}
-          dismissedAlerts={dismissedAlerts}
-          setDismissedAlerts={setDismissedAlerts}
-          page="curriculum"
-        />
+        <AlertsDisplay alerts={alerts} page="curriculum" />
         <div className="col-8">{lessonsToRender}</div>
         <div className="col-4">
           <ProgressCard progressCount={progressPercentage} />
